+++ src/test/java/com/societegenerale/aut/test/ClassTestWithIncorrectName1.java
@@ -1,0 +1,11 @@
+package com.societegenerale.aut.test;
+
+import org.junit.Test;
+
+public class ClassTestWithIncorrectName1 {
+
+	@Test
+	public void check() {
+	}
+
+}