+++ src/test/java/com/societegenerale/aut/test/ClassTestWithIncorrectName2.java
@@ -1,0 +1,17 @@
+package com.societegenerale.aut.test;
+
+import org.junit.Test;
+import org.junit.jupiter.api.Nested;
+
+public class ClassTestWithIncorrectName2 {
+
+	@Nested
+	class PositiveCase {
+
+		@Test
+		public void check() {
+		}
+
+	}
+
+}