+++ src/test/java/com/societegenerale/commons/plugin/rules/ArchUtilsTest.java
@@ -36,7 +36,7 @@ public class ArchUtilsTest {
 
 		long noOfClasses = classes.stream().filter(it -> !it.isNestedClass()).count();
 
-		assertThat(noOfClasses).isEqualTo(30);
+		assertThat(noOfClasses).isEqualTo(32);
 	}
 
 	@Test